+++ dataforge-meta/src/commonMain/kotlin/hep/dataforge/meta/MetaBuilder.kt
@@ -136,7 +136,7 @@ public fun Meta.builder(): MetaBuilder {
 /**
  * Create a deep copy of this meta and apply builder to it
  */
-public fun Meta.edit(builder: MetaBuilder.() -> Unit): MetaBuilder = builder().apply(builder)
+public inline fun Meta.edit(builder: MetaBuilder.() -> Unit): MetaBuilder = builder().apply(builder)
 
 /**
  * Build a [MetaBuilder] using given transformation